refactor: reorganize source folder content
