+++ source/background/background.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-unassigned-import
-import './options-storage.js';
+import '../options/options-storage.js';
