+++ source/content/content.js
@@ -1,4 +1,4 @@
-import optionsStorage from './options-storage.js';
+import optionsStorage from '../options/options-storage.js';
 
 console.log('💈 Content script loaded for', chrome.runtime.getManifest().name);
